Change a href in sidebar to Link, and create Layouts
