+++ src/pages/DetailProject.js
@@ -3,6 +3,7 @@ import Footer from "../components/Footer";
 import {useEffect, useState} from "react";
 import LoadingCard from "../components/LoadingCard";
 import {useParams} from "react-router-dom";
+import MainLayout from "../layouts/MainLayout";
 
 export default function DetailProject() {
     const { id } = useParams();
@@ -107,79 +108,73 @@ export default function DetailProject() {
     }, []);
 
     return (
-        <div className="background h-screen">
-            <div className="mb-10">
-                <Header />
+        <MainLayout>
+            <div className="">
+                {detail.loadStatus === 0 && <LoadingCard />}
 
-                <div className="min-h-[90vh]">
-                    {detail.loadStatus === 0 && <LoadingCard />}
+                {detail.loadStatus === -1 && (
+                    <div className="text-center">
+                        <p style={{ color: 'red' }}>Failed to load project.</p>
+                        <button className="mt-2" onClick={() => loadProject()}>Retry</button>
+                    </div>
+                )}
 
-                    {detail.loadStatus === -1 && (
-                        <div className="text-center">
-                            <p style={{ color: 'red' }}>Failed to load project.</p>
-                            <button className="mt-2" onClick={() => loadProject()}>Retry</button>
-                        </div>
-                    )}
-
-                    {detail.loadStatus === 1 && (
-                        <div className="flex justify-center">
-                            <div className="w-3/4 rounded-lg overflow-hidden shadow-lg my-5 p-4 glass">
-                                <div className="grid grid-flow-row auto-rows-max justify-center gap-5">
-                                    <div className="text-2xl font-extrabold">{ detail.project.title }</div>
+                {detail.loadStatus === 1 && (
+                    <div className="flex justify-center">
+                        <div className="rounded-lg overflow-hidden shadow-lg glass">
+                            <div className="grid grid-flow-row auto-rows-max justify-center gap-5">
+                                <div className="text-2xl font-extrabold">{ detail.project.title }</div>
+                                <div>
+                                    <img src={detail.project.photo} alt="Sunset in the mountains" width="590" height="325"/>
+                                </div>
+                                <button className="text-blue-400 font-bold">See Live Demo</button>
+                                <div className="grid justify-items-start">
+                                    <div className="font-bold">Author :</div>
                                     <div>
-                                        <img src={detail.project.photo} alt="Sunset in the mountains" width="590" height="325"/>
+                                        <a
+                                            href="/profile"
+                                            data-te-placement="right"
+                                            className="transititext-primary text-primary transition duration-150 ease-in-out hover:text-primary-600 focus:text-primary-600 active:text-primary-700 dark:text-primary-400 dark:hover:text-primary-500 dark:focus:text-primary-500 dark:active:text-primary-600"
+                                            data-te-toggle="tooltip"
+                                            title="You can click Author name for more information">
+                                            Mika Handayani (3120550006)
+                                        </a>
                                     </div>
-                                    <button className="text-blue-400 font-bold">See Live Demo</button>
-                                    <div className="grid justify-items-start">
-                                        <div className="font-bold">Author :</div>
-                                        <div>
-                                            <a
-                                                href="/profile"
-                                                data-te-placement="right"
-                                                className="transititext-primary text-primary transition duration-150 ease-in-out hover:text-primary-600 focus:text-primary-600 active:text-primary-700 dark:text-primary-400 dark:hover:text-primary-500 dark:focus:text-primary-500 dark:active:text-primary-600"
-                                                data-te-toggle="tooltip"
-                                                title="You can click Author name for more information">
-                                                Mika Handayani (3120550006)
-                                            </a>
-                                        </div>
-                                        <div className="font-bold mt-2">Lecturers</div>
+                                    <div className="font-bold mt-2">Lecturers</div>
+                                    {
+                                        detail.project.lecturers.map((lecturer, index) => (
+                                            <div key={index}>{lecturer.name} ({lecturer.nip})</div>
+                                        ))
+                                    }
+                                    <div className="font-bold mt-2">Frameworks & Tools</div>
+                                    <div className="flex justify-start">
                                         {
-                                            detail.project.lecturers.map((lecturer, index) => (
-                                                <div key={index}>{lecturer.name} ({lecturer.nip})</div>
+                                            detail.project.technologies.map((tech, index) => (
+                                                <span key={index} className="inline-block bg-white rounded-full border-black border-2 px-3 py-1 text-sm font-semibold mr-2 mb-2">{tech.name}</span>
                                             ))
                                         }
-                                        <div className="font-bold mt-2">Frameworks & Tools</div>
-                                        <div className="flex justify-start">
-                                            {
-                                                detail.project.technologies.map((tech, index) => (
-                                                    <span key={index} className="inline-block bg-white rounded-full border-black border-2 px-3 py-1 text-sm font-semibold mr-2 mb-2">{tech.name}</span>
-                                                ))
-                                            }
-                                        </div>
-                                        <div className="font-bold mt-2">Description</div>
-                                        <div className="grid justify-items-start">
-                                            <div dangerouslySetInnerHTML={{ __html: detail.project.abstract }} />
-                                        </div>
-                                        <div className="font-bold mt-2">Attachments</div>
-                                        <div className="mb-10">
-                                            <a
-                                                href="#"
-                                                data-te-placement="right"
-                                                className="text-blue-600 transititext-primary text-primary transition duration-150 ease-in-out hover:text-primary-600 focus:text-primary-600 active:text-primary-700 dark:text-primary-400 dark:hover:text-primary-500 dark:focus:text-primary-500 dark:active:text-primary-600"
-                                                data-te-toggle="tooltip"
-                                                title="You can click Author name for more information">
-                                                Buku PA
-                                            </a>
-                                        </div>
+                                    </div>
+                                    <div className="font-bold mt-2">Description</div>
+                                    <div className="grid justify-items-start">
+                                        <div dangerouslySetInnerHTML={{ __html: detail.project.abstract }} />
+                                    </div>
+                                    <div className="font-bold mt-2">Attachments</div>
+                                    <div className="mb-10">
+                                        <a
+                                            href="#"
+                                            data-te-placement="right"
+                                            className="text-blue-600 transititext-primary text-primary transition duration-150 ease-in-out hover:text-primary-600 focus:text-primary-600 active:text-primary-700 dark:text-primary-400 dark:hover:text-primary-500 dark:focus:text-primary-500 dark:active:text-primary-600"
+                                            data-te-toggle="tooltip"
+                                            title="You can click Author name for more information">
+                                            Buku PA
+                                        </a>
                                     </div>
                                 </div>
                             </div>
                         </div>
-                    )}
-                </div>
-
-                <Footer />
+                    </div>
+                )}
             </div>
-        </div>
+        </MainLayout>
     )
 }
