+++ src/pages/admin/dashboard-student.js
@@ -4,6 +4,7 @@ import EditStudent from "./dashboard-student/edit-student";
 import DeleteStudent from "./dashboard-student/delete-student";
 import React, {useEffect, useState} from "react";
 import LoadingCard from "../../components/LoadingCard";
+import AdminLayout from "../../layouts/AdminLayout";
 
 export default function DashboardStudent() {
     const [students, setStudent] = useState({
@@ -91,62 +92,55 @@ export default function DashboardStudent() {
     }, [])
 
     return (
-        <div className="flex flex-row">
-            <div className="w-auto bg-blue-400">
-                <Sidebar />
-            </div>
-            <div className="w-full h-screen bg-gray-100">
-                <div className="p-20">
-                    <div className="rounded-sm border border-stroke bg-white px-5 pt-6 pb-2.5 shadow-default dark:border-strokedark dark:bg-boxdark sm:px-7.5 xl:pb-1 ml-40">
-                        <div className="flex justify-between">
-                            <h4 className="mb-6 text-xl font-bold text-black">
-                                Dashboard Student
-                            </h4>
-                            <h1><AddStudent /></h1>
+        <AdminLayout>
+            <div className="rounded-sm border border-stroke bg-white px-5 pt-6 pb-2.5 shadow-default dark:border-strokedark dark:bg-boxdark sm:px-7.5 xl:pb-1 ml-40">
+                <div className="flex justify-between">
+                    <h4 className="mb-6 text-xl font-bold text-black">
+                        Dashboard Student
+                    </h4>
+                    <h1><AddStudent /></h1>
+                </div>
+                <div className="overflow-x-auto">
+                    { students.loadStatus === 0 && <LoadingCard /> }
+                    { students.loadStatus === -1 && (
+                        <div className="text-center">
+                            <p style={{ color: 'red' }}>Failed to load students.</p>
+                            <button className="mt-2" onClick={() => loadStudents()}>Retry</button>
                         </div>
-                        <div className="overflow-x-auto">
-                            { students.loadStatus === 0 && <LoadingCard /> }
-                            { students.loadStatus === -1 && (
-                                <div className="text-center">
-                                    <p style={{ color: 'red' }}>Failed to load students.</p>
-                                    <button className="mt-2" onClick={() => loadStudents()}>Retry</button>
-                                </div>
-                            )}
-                            { students.loadStatus === 1 && (
-                                <table className="min-w-full divide-y divide-gray-200">
-                                    <thead className="bg-gray-300 dark:bg-meta-4">
-                                    <tr>
-                                        <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Name</th>
-                                        <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">NRP</th>
-                                        <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Email</th>
-                                        <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Title</th>
-                                        <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Status</th>
-                                        <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Action</th>
+                    )}
+                    { students.loadStatus === 1 && (
+                        <table className="min-w-full divide-y divide-gray-200">
+                            <thead className="bg-gray-300 dark:bg-meta-4">
+                            <tr>
+                                <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Name</th>
+                                <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">NRP</th>
+                                <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Email</th>
+                                <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Title</th>
+                                <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Status</th>
+                                <th className="p-2.5 xl:p-5 text-sm font-medium uppercase text-center">Action</th>
+                            </tr>
+                            </thead>
+                            <tbody className="divide-y divide-stroke dark:divide-strokedark">
+                            {
+                                students.students.map((d, i) => (
+                                    <tr key={i}>
+                                        <td className="p-2.5 xl:p-5">{d.name}</td>
+                                        <td className="p-2.5 xl:p-5">{d.nrp}</td>
+                                        <td className="p-2.5 xl:p-5">{d.email}</td>
+                                        <td className="p-2.5 xl:p-5">{d.title}</td>
+                                        <td className={"p-2.5 xl:p-5 " + (d.status === 'Inactive' ? 'text-red-600' : '')}>{d.status}</td>
+                                        <td className="p-2.5 xl:p-5">
+                                            <EditStudent />
+                                            <DeleteStudent />
+                                        </td>
                                     </tr>
-                                    </thead>
-                                    <tbody className="divide-y divide-stroke dark:divide-strokedark">
-                                    {
-                                        students.students.map((d, i) => (
-                                            <tr key={i}>
-                                                <td className="p-2.5 xl:p-5">{d.name}</td>
-                                                <td className="p-2.5 xl:p-5">{d.nrp}</td>
-                                                <td className="p-2.5 xl:p-5">{d.email}</td>
-                                                <td className="p-2.5 xl:p-5">{d.title}</td>
-                                                <td className={"p-2.5 xl:p-5 " + (d.status === 'Inactive' ? 'text-red-600' : '')}>{d.status}</td>
-                                                <td className="p-2.5 xl:p-5">
-                                                    <EditStudent />
-                                                    <DeleteStudent />
-                                                </td>
-                                            </tr>
-                                        ))
-                                    }
-                                    </tbody>
-                                </table>
-                            )}
-                        </div>
-                    </div>
+                                ))
+                            }
+                            </tbody>
+                        </table>
+                    )}
                 </div>
             </div>
-        </div>
+        </AdminLayout>
     )
 }
